# content/schemas/openapi.json
{
  "components": {
    "schemas": {
      "ActivityListResponse": {
        "additionalProperties": false,
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/ActivityView"
            },
            "title": "Members",
            "type": "array"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id"
        ],
        "title": "ActivityListResponse",
        "type": "object"
      },
      "ActivityStartRequest": {
        "additionalProperties": false,
        "properties": {
          "character_id": {
            "format": "uuid",
            "title": "Character Id",
            "type": "string"
          },
          "duration_phases": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Phases"
          },
          "kind": {
            "maxLength": 32,
            "title": "Kind",
            "type": "string"
          },
          "skill": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill"
          },
          "to_location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Location Id"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id",
          "character_id",
          "kind"
        ],
        "title": "ActivityStartRequest",
        "type": "object"
      },
      "ActivityView": {
        "additionalProperties": false,
        "properties": {
          "character_id": {
            "format": "uuid",
            "title": "Character Id",
            "type": "string"
          },
          "duration_phases": {
            "title": "Duration Phases",
            "type": "integer"
          },
          "effective_progress_phases": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Progress Phases"
          },
          "from_location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Location Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress_phases": {
            "title": "Progress Phases",
            "type": "integer"
          },
          "route_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id"
          },
          "start_absolute": {
            "title": "Start Absolute",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "to_location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Location Id"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "world_id",
          "character_id",
          "kind",
          "status",
          "start_absolute",
          "duration_phases",
          "progress_phases",
          "version"
        ],
        "title": "ActivityView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "worldsim",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/stage2/activities": {
      "get": {
        "description": "Active activities for a world, oldest first.",
        "operationId": "list_activities_api_v1_stage2_activities_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Activities",
        "tags": [
          "activities"
        ]
      },
      "post": {
        "description": "Begin one activity for a character (one active at a time).",
        "operationId": "start_api_v1_stage2_activities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/stage2/activities/{activity_id}/cancel": {
      "post": {
        "description": "Terminally abandon an activity.",
        "operationId": "cancel_api_v1_stage2_activities__activity_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/stage2/activities/{activity_id}/interrupt": {
      "post": {
        "description": "Freeze an active activity, baking elapsed progress.",
        "operationId": "interrupt_api_v1_stage2_activities__activity_id__interrupt_post",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interrupt",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/stage2/activities/{activity_id}/resume": {
      "post": {
        "description": "Restart an interrupted activity from the current phase.",
        "operationId": "resume_api_v1_stage2_activities__activity_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume",
        "tags": [
          "activities"
        ]
      }
    },
...
  }
}